+++ src/components/mypage/StudyList.jsx
@@ -4,13 +4,12 @@ import { Color } from '../style/Color';
 import RightArrow from '../../assets/icons/rightarrow.svg?react';
 import studyProfileUrl from '../../assets/images/studyprofile.png';
 
-const CurrentStudyList = () => {
+const StudyList = ({ isCurrent }) => {
     return (
         <StudyListWrapper>
             <Wrapper>
                 <RowWrapper>
-                    <ExtraBold>현재 스터티룸</ExtraBold>
-                    <RowWrapper2><GreyText>모두 보기</GreyText><StyledRightArrow/></RowWrapper2>
+                    <ExtraBold>{isCurrent ? "현재 스터디룸" : "이전 스터디룸"}</ExtraBold>
                 </RowWrapper>
                 <ListWrapper>
                     <ListItem>
@@ -40,7 +39,7 @@ const CurrentStudyList = () => {
     );
 };
 
-export default CurrentStudyList;
+export default StudyList;
 
 export const StyledRightArrow = styled(RightArrow)`
     width : 0.495em;
@@ -50,7 +49,7 @@ export const StudyListWrapper = styled.div`
     width : 100%;
     display : flex;
     gap :2em;
-`
+`;
 
 export const Wrapper = styled.div`
     width : 100%;
@@ -66,25 +65,20 @@ export const ListWrapper = styled.div`
     display : flex;
     flex-direction : column;
     gap : 1em;
-`
+`;
 
 export const ListItem = styled.div`
     display : flex;
     gap : 1em;
     border-bottom : 1px solid #8E59FF;
     padding-bottom : 1.25em;
-`
+`;
 
 export const RowWrapper = styled.div`
     width : 100%;
     display : flex;
     align-items : center;
     justify-content: space-between;
-`
-export const RowWrapper2 = styled.div`
-    display : flex;
-    align-items : center;
-    gap : 0.5em;
 `;
 
 export const ColumnWrapper = styled.div`
@@ -94,12 +88,6 @@ export const ColumnWrapper = styled.div`
     gap : 0.5em;
 `;
 
-export const GreyText = styled.div`
-    color : #D0D1D9;
-    font-size : 1em;
-    font-weight : 800;
-`;
-
 export const ExtraBold = styled(Color)`
     font-weight : 800;
     font-size : 1.25em;
@@ -107,8 +95,8 @@ export const ExtraBold = styled(Color)`
 
 export const StudyName = styled.div`
     font-size : 1em;
-    font-weight : 700
-`
+    font-weight : 700;
+`;
 
 export const StudyText = styled.div`
     font-size : 1em;
